requirement.txt
absl-py==0.13.0
asgiref==3.4.1
Django==3.2.7
djangorestframework==3.12.4
numpy==1.21.2
ortools==9.0.9048
protobuf==3.17.3
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
